bugfix: secondary buttons no longer perform form submission
the default value of the "type" attribute of the HTML "button" element
is "submit". secondary buttons must therefore have their "type"
attribute explicitly set to "button" to prevent form submission. while
we're at it, we also explicitly set the primary buttons' type so that
in the future we won't run into the same problem when we copy&paste a
snippet.