all links are now based on an absolute URL base path
the reason for this change is that for the upcoming AngularJS
implementation of showing the game board we want to have URLs of
depth 2, i.e. "board/42". up until now we always had URLs of depth 1.
handling URLs of varying depth is at least difficult, if not outright
impossible, to achieve with links that have relative URLs. basing all
links on an absolute path is much, much simpler.