shift responsibility for resigning game from modal controller to calling controller
this is a simple solution for the problem that the calling controller
must somehow get a feedback from the modal whether the user actually
did resign the game. multiple other solutions proved to be much more
complicated, but without any actual additional benefit except vague
promises of "better decoupling" or "more flexible". this new solution
is straight to the point and is sufficient for now. if better decoupling
or more flexibility is needed it can be added later.